refactored main.dart to separate files
