+++ lib/pages/start_page.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class StartPage extends StatefulWidget {
+  StartPage({Key key, this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  _StartPageState createState() => _StartPageState();
+}
+
+class _StartPageState extends State<StartPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          title: Text(widget.title),
+        ),
+        body: Container());
+  }
+}
